first code from views to utils and models
